change NO_REAL_FD to u64

+++ src/example_grates/commonconstants.rs
@@ -5,7 +5,7 @@
 pub const FD_PER_PROCESS_MAX: i32 = 1024;
 
 /// Use this to indicate there isn't a real fd backing an item
-pub const NO_REAL_FD: i32 = 0xffabcdef01;
+pub const NO_REAL_FD: u64 = 0xffabcdef01;
 
 // BUG / TODO: Use this in some sane way...
 #[allow(dead_code)]
